Fix ValueAtPoints data type
max_values_in_array should be an integer not a double.

Also update schema files.

+++ include/world_builder/types/value_at_points.h
@@ -62,7 +62,7 @@ namespace WorldBuilder
                           const std::string &documentation) const override final;
 
         double default_value;
-        double max_values_in_array;
+        size_t max_values_in_array;
         std::vector<Point<2> > default_points;
 
       protected:
